Add files via upload
LDR_bases sensing
Ldr fs used to take input from Light Depended Resistor to impart day and night sense on the system
The infrared sensor is used to estimate the distance of the incoming object which in turns allows the Arduino board to switch on/off connected street lights.

Sonar and Vibration bases sensing
Infrared can't work properly in daylight as the sun's emitted infrared interfere with the onboard infrared sensor. To overcome this Sonar and Underground vibration sensor 
are deployed.